+++ src/engine/util/hash.h
@@ -1,8 +1,9 @@
 #ifndef HASH_H
 #define HASH_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-uint64_t str_hash(const void* v);
+size_t str_hash(const void* v);
 
 #endif
